+++ src/Callbacks/AbstractSelectCallback.php
@@ -23,7 +23,6 @@ namespace TechDivision\Import\Callbacks;
 use TechDivision\Import\Utils\MemberNames;
 use TechDivision\Import\Utils\RegistryKeys;
 use TechDivision\Import\Utils\StoreViewCodes;
-use TechDivision\Import\Utils\FrontendInputTypes;
 
 /**
  * A callback implementation that converts the passed select value.
@@ -34,7 +33,7 @@ use TechDivision\Import\Utils\FrontendInputTypes;
  * @link      https://github.com/techdivision/import
  * @link      http://www.techdivision.com
  */
-class SelectCallback extends AbstractCallback
+abstract class AbstractSelectCallback extends AbstractCallback
 {
 
     /**
@@ -75,7 +74,10 @@ class SelectCallback extends AbstractCallback
                 array(
                     RegistryKeys::MISSING_OPTION_VALUES => array(
                         $attributeCode => array(
-                            $attributeValue => FrontendInputTypes::SELECT
+                            $attributeValue => array(
+                                $this->raiseCounter($attributeValue),
+                                array($this->getUniqueIdentifier() => true)
+                            )
                         )
                     )
                 )