+++ src/Plugins/MissingOptionValuesPlugin.php
@@ -85,14 +85,17 @@ class MissingOptionValuesPlugin extends AbstractPlugin
                 ColumnKeys::STORE_VIEW_CODE,
                 ColumnKeys::ATTRIBUTE_CODE,
                 ColumnKeys::VALUE,
+                ColumnKeys::COUNTER,
+                ColumnKeys::UNIQUE_IDENTIFIER,
                 ColumnKeys::SORT_ORDER
             )
         );
 
         // append the missing option values to the array
         foreach ($missingOptions as $attributeCode => $options) {
-            foreach (array_keys($options) as $value) {
-                $toBeCreated[] = array(null, $attributeCode, $value, null);
+            foreach ($options as $value => $data) {
+                list($counter, $skus) = $data;
+                $toBeCreated[] = array(null, $attributeCode, $value, $counter, implode(',', array_keys($skus)), null);
             }
         }