+++ src/Utils/ColumnKeys.php
@@ -132,4 +132,18 @@ class ColumnKeys
      * @var string
      */
     const SORT_ORDER = 'sort_order';
+
+    /**
+     * Name for the column 'counter'.
+     *
+     * @var string
+     */
+    const COUNTER = 'counter';
+
+    /**
+     * Name for the column 'uniqueIdentifier'.
+     *
+     * @var string
+     */
+    const UNIQUE_IDENTIFIER = 'uniqueIdentifier';
 }