+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 1.0.0-alpha49
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Make select, multiselect + boolean callbacks abstract
+* Refactoring for DI integration
+
 # Version 1.0.0-alpha48
 
 ## Bugfixes